arc: move hooks-warmer DaemonSet from arc-runners to arc module
**Impact:** clusters with `arc` but without `arc-runners` (e.g. UW1)
**Risk:** low

## What
Relocates the `hooks-warmer` DaemonSet and its smoke tests from `modules/arc-runners/` to `modules/arc/`, so clusters that enable `arc` without `arc-runners` still get the hooks-warmer deployed.

## Why
The UW1 cluster runs `arc` but not `arc-runners`. The hooks-warmer DaemonSet — which pre-downloads patched runner-container-hooks onto c7i-runner nodes — was incorrectly scoped to `arc-runners`, making it unavailable on UW1. It belongs to the `arc` module because it is a prerequisite for any runner pod, not specific to the runner scale-set definitions managed by `arc-runners`.

## How
- Used `git mv` to preserve file history on `hooks-warmer.yaml`
- Moved the smoke test class into a new dedicated test file under `modules/arc/tests/smoke/` (the test is independent of runner-def machinery and checks `enabled_modules` for `arc` instead of `arc-runners`)
- Removed the now-empty `arc-runners/kubernetes/` directory (namespace.yaml and kustomization.yaml deleted — the namespace is already managed by the `arc` module)

## Changes
- `modules/arc-runners/kubernetes/hooks-warmer.yaml` → `modules/arc/kubernetes/hooks-warmer.yaml` (git mv, no content change)
- `modules/arc/kubernetes/kustomization.yaml`: added `hooks-warmer.yaml` to the resource list
- `modules/arc-runners/kubernetes/kustomization.yaml`: deleted (no remaining resources)
- `modules/arc-runners/kubernetes/namespace.yaml`: deleted (arc-runners namespace already created by arc module)
- `modules/arc-runners/tests/smoke/test_runners.py`: removed `TestHooksWarmer` class and unused imports (`assert_daemonset_healthy`, `filter_daemonsets`)
- `modules/arc/tests/smoke/test_hooks_warmer.py`: new file with `TestHooksWarmer` class, gated on `arc` module instead of `arc-runners`
- `docs/modules.md`: moved hooks-warmer docs from `arc-runners` to `arc` section
- `docs/node-warmup-and-scheduling-gates.md`: updated file path references for hooks-warmer

## Testing
- `just lint` — all linters pass
- `just test` — unit tests pass
- Smoke tests: run `pytest modules/arc/tests/smoke/test_hooks_warmer.py` against a cluster with `arc` enabled to verify the DaemonSet is detected and healthy
- Verify on UW1: deploy `arc` module and confirm hooks-warmer DaemonSet is created in `arc-runners` namespace
