--- osdc/modules/arc-runners/kubernetes/namespace.yaml
@@ -1,6 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: arc-runners
-  labels:
-    app.kubernetes.io/part-of: osdc-arc-runners
