+++ osdc/modules/arc/kubernetes/hooks-warmer.yaml
