+++ osdc/modules/arc/kubernetes/kustomization.yaml
@@ -1,7 +1,7 @@
 # ARC module — Kubernetes resources for Actions Runner Controller.
 #
 # Applied when 'arc' is in the cluster's module list.
-# Deploys: namespaces, runner serviceaccount, limitrange, hooks configmap.
+# Deploys: namespaces, runner serviceaccount, limitrange, hooks configmap, hooks-warmer DaemonSet.
 # The ARC controller Helm chart is installed by deploy.sh (not kustomize).
 # Registry mirror config is in base/kubernetes/ (all clusters need it, not just ARC).
 apiVersion: kustomize.config.k8s.io/v1beta1
@@ -12,3 +12,4 @@ resources:
   - runner-serviceaccount.yaml
   - runner-limitrange.yaml
   - runner-hooks-configmap.yaml
+  - hooks-warmer.yaml
